refactor: extract NewTraining sections
