+++ frontend/src/components/Stockbot/NewTraining/utils.ts
@@ -1,0 +1,4 @@
+export const safeNum = (v: any, fallback = 0) => {
+  const n = typeof v === "string" ? parseFloat(v) : v;
+  return Number.isFinite(n) ? n : fallback;
+};
